feat: event system with notification to client
* feat: user name change api, split user name to first and last

* feat: organization email -> admins

* feat: event system

* feat: Gateway separation, align with event system

* feat: event emit for Team

* feat: refresh signal to client

* fix: duplicate user gateway instantiation

* feat: split chatroom, refresh request

+++ src/modules/team/team.service.ts
@@ -12,11 +12,13 @@ import { CreateTeamDto } from './dto/create-team.dto';
 import { Membership } from './membership.entity';
 import { JoinTeamDto } from './dto/join-team.dto';
 import { ChatroomParticipant } from '../chatroom/chatroom-participant.entity';
+import { EventManager } from 'src/common/events/event-manager';
 
 @Injectable()
 export class TeamService {
 	constructor(
 		private readonly dataSource: DataSource,
+		private readonly eventManager: EventManager,
 
 		@InjectRepository(Team)
 		private readonly teamRepo: Repository<Team>,
@@ -53,6 +55,19 @@ export class TeamService {
 					role: 'admin',
 				});
 				await manager.getRepository(Membership).save(membership);
+
+				const memberships = await manager.getRepository(Membership).find({
+					where: { organization_id: dto.organizationId },
+					select: ['user_id'],
+				});
+				const userIds = memberships.map((m) => m.user_id);
+				if (userIds.length > 0) {
+					this.eventManager.emit('user.data_dirty', { userIds });
+				}
+				this.eventManager.emit('user.chatroom_join', {
+					userId: user_id,
+					chatroomIds: [savedChatroom.id],
+				});
 
 				const participant = manager.getRepository(ChatroomParticipant).create({
 					user_id,
@@ -103,15 +118,43 @@ export class TeamService {
 			});
 		}
 
+		const members = await this.membershipRepo.find({
+			where: { team_id: team.id },
+			select: ['user_id'],
+		});
+
+		const allTeamMemberIds = members.map((m) => m.user_id);
+
+		this.eventManager.emit('user.data_dirty', {
+			userIds: allTeamMemberIds,
+		});
+		if (team.chatroom_id) {
+			this.eventManager.emit('user.chatroom_join', {
+				userId: user_id,
+				chatroomIds: [team.chatroom_id],
+			});
+		}
+
 		return { success: true };
 	}
 
 	async delete(id: number) {
-		await this.membershipRepo.delete({ team_id: id });
+		const team = await this.teamRepo.findOneBy({ id });
+		if (!team) throw new NotFoundException('Team not found');
 
-		const result = await this.teamRepo.delete(id);
-		if (result.affected === 0) {
-			throw new NotFoundException('Team not found');
+		const memberships = await this.membershipRepo.find({
+			where: { team_id: id },
+			select: ['user_id'],
+		});
+
+		const userIds = memberships.map((m) => m.user_id);
+
+		await this.membershipRepo.delete({ team_id: id });
+		await this.teamRepo.delete(id);
+
+		if (userIds.length > 0) {
+			// TODO: Socket close
+			this.eventManager.emit('user.data_dirty', { userIds });
 		}
 
 		return { success: true };
